requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-django==4.9.0
python-decouple==3.8
